+++ src/test/scala/com/madewithtea/mockedstreams/MockedStreamsSpec.scala
@@ -32,52 +32,38 @@ import org.apache.kafka.streams.scala.StreamsBuilder
 import org.apache.kafka.streams.scala.kstream.KTable
 import org.apache.kafka.streams.state.ValueAndTimestamp
 import org.scalatest.{FlatSpec, Matchers}
+import java.time.Duration
+import org.apache.kafka.streams.processor.ProcessorContext
+import com.madewithtea.mockedstreams.MockedStreams.TopologyNotSet
 
 class MockedStreamsSpec extends FlatSpec with Matchers {
   import CustomEquality._
 
   behavior of "MockedStreams"
 
-  it should "throw exception when expected size in output methods is <= 0" in {
-    import Fixtures.Uppercase._
-    import MockedStreams.ExpectedOutputIsEmpty
-
-    val spec = MockedStreams()
-      .topology(topology)
-      .input(InputTopic, strings, strings, input)
-
-    Seq(-1, 0).foreach { size =>
-      an[ExpectedOutputIsEmpty] should be thrownBy
-        spec.output(OutputTopic, strings, strings, size)
-
-      an[ExpectedOutputIsEmpty] should be thrownBy
-        spec.outputTable(OutputTopic, strings, strings, size)
-    }
+  it should "throw exception when inputs specified before topology" in {
+    an[TopologyNotSet] should be thrownBy
+      MockedStreams().input("input", stringSerde, stringSerde, Seq())
   }
 
-  it should "throw exception when no input specified for all output and state methods" in {
-    import Fixtures.Uppercase._
-    import MockedStreams.NoInputSpecified
+  it should "throw exception when outputs specified before topology" in {
+    an[TopologyNotSet] should be thrownBy
+      MockedStreams().output("output", stringSerde, stringSerde)
+  }
 
-    val t = MockedStreams().topology(topology)
+  it should "throw exception when outputTable specified before topology" in {
+    an[TopologyNotSet] should be thrownBy
+      MockedStreams().outputTable("output", stringSerde, stringSerde)
+  }
 
-    an[NoInputSpecified] should be thrownBy
-      t.output(OutputTopic, strings, strings, expected.size)
-
-    an[NoInputSpecified] should be thrownBy
-      t.outputTable(OutputTopic, strings, strings, expected.size)
-
-    an[NoInputSpecified] should be thrownBy
-      t.stateTable("state-table")
-
-    an[NoInputSpecified] should be thrownBy
-      t.windowStateTable("window-state-table", 0)
-
-    an[NoInputSpecified] should be thrownBy
-      t.windowStateTable("window-state-table",
-                         0,
-                         Instant.ofEpochMilli(Long.MinValue),
-                         Instant.ofEpochMilli(Long.MaxValue))
+  it should "throw exception state store access before topology " in {
+    an[TopologyNotSet] should be thrownBy
+      MockedStreams().windowStateTable(
+        "table",
+        "key",
+        Instant.now(),
+        Instant.now().plusMillis(1)
+      )
   }
 
   it should "assert correctly when processing strings to uppercase" in {
@@ -86,7 +72,7 @@ class MockedStreamsSpec extends FlatSpec with Matchers {
     val output = MockedStreams()
       .topology(topology)
       .input(InputTopic, strings, strings, input)
-      .output(OutputTopic, strings, strings, expected.size)
+      .output(OutputTopic, strings, strings)
 
     output shouldEqual expected
   }
@@ -97,7 +83,7 @@ class MockedStreamsSpec extends FlatSpec with Matchers {
     val output = MockedStreams()
       .topology(topology)
       .input(InputTopic, strings, strings, input)
-      .outputTable(OutputTopic, strings, strings, expected.size)
+      .outputTable(OutputTopic, strings, strings)
 
     output shouldEqual expected.toMap
   }
@@ -111,7 +97,7 @@ class MockedStreamsSpec extends FlatSpec with Matchers {
       .input(InputBTopic, strings, ints, inputB)
       .stores(Seq(StoreName))
 
-    builder.output(OutputATopic, strings, ints, expectedA.size) shouldEqual expectedA
+    builder.output(OutputATopic, strings, ints) shouldEqual expectedA
     builder.stateTable(StoreName) shouldEqual inputA.toMap
   }
 
@@ -125,11 +111,11 @@ class MockedStreamsSpec extends FlatSpec with Matchers {
       .stores(Seq(StoreName))
 
     builder
-      .output(OutputATopic, strings, ints, expectedA.size)
+      .output(OutputATopic, strings, ints)
       .shouldEqual(expectedA)
 
     builder
-      .output(OutputBTopic, strings, ints, expectedB.size)
+      .output(OutputBTopic, strings, ints)
       .shouldEqual(expectedB)
 
     builder.stateTable(StoreName) shouldEqual inputA.toMap
@@ -151,7 +137,7 @@ class MockedStreamsSpec extends FlatSpec with Matchers {
       .input(InputATopic, strings, ints, secondInputForTopicA)
 
     builder
-      .output(OutputATopic, strings, ints, expectedOutput.size)
+      .output(OutputATopic, strings, ints)
       .shouldEqual(expectedOutput)
   }
 
@@ -161,8 +147,10 @@ class MockedStreamsSpec extends FlatSpec with Matchers {
     import Fixtures.Multi._
 
     val props = new Properties
-    props.put(StreamsConfig.DEFAULT_TIMESTAMP_EXTRACTOR_CLASS_CONFIG,
-              classOf[TimestampExtractors.CustomTimestampExtractor].getName)
+    props.put(
+      StreamsConfig.DEFAULT_TIMESTAMP_EXTRACTOR_CLASS_CONFIG,
+      classOf[TimestampExtractors.CustomTimestampExtractor].getName
+    )
 
     val builder = MockedStreams()
       .topology(topology1WindowOutput)
@@ -177,21 +165,24 @@ class MockedStreamsSpec extends FlatSpec with Matchers {
     builder
       .windowStateTable[String, Int](StoreName, "y")
       .shouldEqual(expectedCy.toMap)
-      
-    builder
-      .windowStateTable[String, Int](StoreName,
-                                      "y",
-                                      Instant.ofEpochMilli(0L),
-                                      Instant.ofEpochMilli(1L))
-      .shouldEqual(expectedCy.toMap)
-      
 
     builder
-      .windowStateTable[String, Int](StoreName,
-                                      "x",
-                                      Instant.ofEpochMilli(0L),
-                                      Instant.ofEpochMilli(1L))
-     .shouldEqual(expectedCx.toMap)
+      .windowStateTable[String, Int](
+        StoreName,
+        "y",
+        Instant.ofEpochMilli(0L),
+        Instant.ofEpochMilli(1L)
+      )
+      .shouldEqual(expectedCy.toMap)
+
+    builder
+      .windowStateTable[String, Int](
+        StoreName,
+        "x",
+        Instant.ofEpochMilli(0L),
+        Instant.ofEpochMilli(1L)
+      )
+      .shouldEqual(expectedCx.toMap)
   }
 
   it should "accept already built topology" in {
@@ -206,13 +197,12 @@ class MockedStreamsSpec extends FlatSpec with Matchers {
     val output = MockedStreams()
       .withTopology(() => getTopology)
       .input(InputTopic, strings, strings, input)
-      .output(OutputTopic, strings, strings, expected.size)
+      .output(OutputTopic, strings, strings)
 
     output shouldEqual expected
   }
 
   it should "accept consumer records with custom timestamps" in {
-
     import Fixtures.Multi._
 
     val builder = MockedStreams()
@@ -225,10 +215,12 @@ class MockedStreamsSpec extends FlatSpec with Matchers {
       .shouldEqual(expectedCWithTimeStamps.toMap)(valueAndTimestampEq[Int])
 
     builder
-      .windowStateTable[String, Long](StoreName,
-                        "x",
-                        Instant.ofEpochMilli(1000L),
-                        Instant.ofEpochMilli(1002L))
+      .windowStateTable[String, Long](
+        StoreName,
+        "x",
+        Instant.ofEpochMilli(1000L),
+        Instant.ofEpochMilli(1002L)
+      )
       .shouldEqual(expectedCWithTimeStamps.toMap)
   }
 
@@ -275,8 +267,10 @@ class MockedStreamsSpec extends FlatSpec with Matchers {
       val expectedA = Seq(("x", 5), ("y", 5))
       val expectedB = Seq(("x", 3), ("y", 1))
 
-      val expectedCx = Seq((1L, ValueAndTimestamp.make(2, 1L)),
-                           (2L, ValueAndTimestamp.make(1, 2L)))
+      val expectedCx = Seq(
+        (1L, ValueAndTimestamp.make(2, 1L)),
+        (2L, ValueAndTimestamp.make(1, 2L))
+      )
       val expectedCy = Seq((1, 1))
 
       val expectedCWithTimeStamps = Seq(
@@ -316,7 +310,7 @@ class MockedStreamsSpec extends FlatSpec with Matchers {
       def topology1WindowOutput(builder: StreamsBuilder) = {
         val streamA = builder.stream[String, Int](InputCTopic)
         streamA.groupByKey
-          .windowedBy(TimeWindows.of(1))
+          .windowedBy(TimeWindows.of(Duration.ofMillis(1)))
           .count()(Materialized.as(StoreName))
       }
 
@@ -364,8 +358,10 @@ class MockedStreamsSpec extends FlatSpec with Matchers {
 
 object TimestampExtractors {
   class CustomTimestampExtractor extends TimestampExtractor {
-    override def extract(record: ConsumerRecord[AnyRef, AnyRef],
-                         previous: Long): Long = record.value match {
+    override def extract(
+        record: ConsumerRecord[AnyRef, AnyRef],
+        previous: Long
+    ): Long = record.value match {
       case value: Integer => value.toLong
       case _              => record.timestamp()
     }
@@ -375,9 +371,13 @@ object TimestampExtractors {
 object CustomEquality {
   import org.scalactic.Equality
 
-  implicit def valueAndTimestampEq[A]: Equality[Map[java.lang.Long, ValueAndTimestamp[A]]] = 
-  new Equality[Map[java.lang.Long, ValueAndTimestamp[A]]] {
-      override def areEqual(a: Map[java.lang.Long, ValueAndTimestamp[A]], b: Any): Boolean = {
+  implicit def valueAndTimestampEq[A]
+      : Equality[Map[java.lang.Long, ValueAndTimestamp[A]]] =
+    new Equality[Map[java.lang.Long, ValueAndTimestamp[A]]] {
+      override def areEqual(
+          a: Map[java.lang.Long, ValueAndTimestamp[A]],
+          b: Any
+      ): Boolean = {
         true
       }
     }